+++ app/src/main/java/com/jeremyhahn/cropdroid/ui/condition/ConditionViewModelFactory.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.jeremyhahn.cropdroid.data.CropDroidAPI
 
-class ConditionViewModelFactory(private val cropDroidAPI: CropDroidAPI, private val channelId: Int) : ViewModelProvider.Factory {
+class ConditionViewModelFactory(private val cropDroidAPI: CropDroidAPI, private val channelId: Long) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return ConditionViewModel(cropDroidAPI, channelId) as T
     }
